Add handout target, fix "always build" issue
This adds a "handout" target which will build the presentation with a
new, bogus command \HANDOUT defined which will turn on a handout build,
i.e., all overlays are compressed down to a single slide. This may
require special definition of \only like \only<2|handout:0> to disable
certain reveals for the handout.

This also fixes the kludginess related to getting the Makefile to always
build the paper and presentation regardless of what the Makefile thinks
of the dependencies. This is necessary as the Makefile cannot figure out
dependencies on its own, but has to rely on latexmk. To do this, I use a
phony rule "go-for-it", the old Tom Cruise standby.